fix size check when uploading chunks
The current size must be calculated from existing chunks and not
file-metadata. Otherwise when doing only 1 simultaneous upload, the
next chunk is checked against complete file size and not existing
chunks.

This also means that it is not possible to strictly check size limits
when doing multiple simultaneous uploads, since they may arrive “at
the same time” and all are checked against the current state but not
against each other.

Refs: #13