Allow to map database error messages to user messages
When adding more checks to the database, the error messages are
usually not very helpful to the end user. An admin is now able to
translate them into better messages.

Concrete use described in #255, where a virus-check is performed
against uploaded files. The failure can be transported to the user in
a better message.

Additionally, these errors are interpreted as "permanent" and thus are
not retried from the web client. The web client knows about this
response code and extracts the message from the tus error message.

Issue: #255