Use a dynamic base-url if the base-url setting is unchanged
To better support access from other machines and for improving startup
experience, the base-url is determined using the request if the
setting is left at its default value (i.e. if the address is
`localhost`). If the `base-url` setting is changed, the behavior is as
before – the base-url is send to the client which uses it to create
requests.

The domain attribute of the cookie is removed, as the default behavior
is better as it excludes subdomains. The cookie will be only send to
the exact same origin.