Fix range responses
Use a fixed chunk-size in case the request specifies an open range
which would mean to transport the entire file. In this case, the
client gets some smaller chunk as it is already indicating to
understand range responses. If a file is smaller than this chunksize,
it will be a normal response.

Fixes: #1328